# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HART Estate API",
        "version": "0.0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Войти в систему",
                "description": "Войти в систему",
                "operationId": "App\\Http\\Controllers\\Api\\Auth\\AuthController::login",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}